refactor: 디스코드 알림 구조 개선
