+++ dongsan-external-api/src/main/java/com/dongsan/api/support/response/discord/DiscordNotifier.java
@@ -1,0 +1,9 @@
+package com.dongsan.api.support.response.discord;
+
+import jakarta.servlet.http.HttpServletRequest;
+import org.springframework.stereotype.Component;
+
+@Component
+public interface DiscordNotifier {
+    void notify(Exception e, HttpServletRequest request);
+}
